requirements.txt
functions_framework
google-cloud
google-cloud-bigquery
google-cloud-storage
pandas
fsspec
gcsfs
pyarrow